fix: critical kernel correctness, memory safety, and robustness across all backends
Q3_K dequantization (all 6 backends):
- Scale decoding was using only 4-bit nibbles (bias -8) when the format
  stores 6-bit values (lo4 | hi2<<4, bias -32). Bytes 8-11 of
  scales_raw hold the high 2 bits, previously ignored. Fixes incorrect
  output for all Q3_K quantized models on CPU, CUDA, Metal, ROCm,
  Vulkan, and WebGPU.

Megakernel atomic float reduction (CUDA, ROCm, Metal):
- RMS norm global sum used atom.global.add.u32 on bitcast float values —
  integer addition of IEEE 754 bit patterns is NOT float addition.
  CUDA/ROCm: changed to atom.global.add.f32. Metal: replaced with CAS
  loop for correct float atomic add. Affects Gemma Q4K/Q8 and Qwen 3.5
  Q8 megakernels.

Metal megakernel grid sync:
- Sync counter was not phase-aware: back-to-back syncs (e.g. inside
  mega_rms_norm) could race because both checked for the same target.
  Now auto-detects phase from counter value: target = (val/n_tgs + 1)
  * n_tgs.

SDPA tree kernel (CUDA, ROCm):
- score_buf was per-thread local memory but read by all threads in the
  V-accumulate loop — data race. Moved to shared memory with explicit
  offsets (sc_off, bcast_off), proper syncthreads barriers, and
  broadcast slots for reduced values.

Metal SDPA fixes:
- Q vector load used strided access (tid, tid+tg_size, ...) into
  thread-local q_local[] but later read all elements — uninitialized
  gaps. Fixed to sequential full load.
- Division by l_i replaced with multiply-by-reciprocal with zero guard
  to prevent NaN output when all attention weights are zero.

clampedSiluMul (all backends):
- Gate value was only clamped above (min(gate, 10)) but not below,
  allowing -inf to reach exp() and produce NaN. Now symmetric [-10, 10]
  clamp on both gate and up. CUDA version adds flushActivations/
  invalidateAct for correct GPU↔CPU coherence.

Softmax NaN on grammar-masked inputs:
- All-negative-infinity logits (from grammar masking) produce
  -inf - (-inf) = NaN through exp, propagating NaN to output. Now
  detects non-finite or zero sum and zeros output. Added tests for
  all-inf and partial-inf cases.

CUDA backend safety:
- Activation cache: stale entries not removed after cuMemFree, causing
  use-after-free. Now removes cache entry on size mismatch.
- Memory leak: failed cuMemcpyHtoD did not free the allocated device
  pointer. Added cleanup on error path.
- SDPA shared memory: long contexts could exceed 48KB smem limit,
  silently producing garbage. Now checks against max_smem_bytes (49152)
  and panics with diagnostic message.
- Fused FFN: missing quant type silently fell through. Now explicit
  @panic on unsupported types.
- KV buffer: null pointer after failed cuMemAlloc not caught.
- uploadToDevice: failures now @panic instead of continuing with
  uninitialized device memory.

Vulkan shared memory (all ~30 shaders):
- shared_sums[8] assumed subgroup size 32 (256/32=8 slots). GPUs with
  subgroup size 8 need 32 slots. Fixed to shared_sums[32] across all
  GEMV, norm, softmax, and SDPA shaders.

Vulkan Q3_K shader:
- Same 6-bit scale fix as other backends: extract hi2 bits from
  bytes 8-11 and combine with lo4.

RadixTree hash collision:
- Hash cache entries lacked query_len, so different token sequences
  with colliding xxHash64 values could return wrong match results.
  Added query_len to HashCacheEntry and verify on lookup.
- Linear probe in insert could overwrite existing child on probe
  exhaustion. Now frees leaf node instead of corrupting tree.

GGUF multi-shard memory leaks:
- Shard parse failure leaked owned_strings, owned_arrays, and
  owned_u32_arrays. Added cleanup in error path.
- Shard resource transfer used catch-and-ignore on appendSlice,
  silently leaking already-referenced heap allocations. Now propagates
  OOM with try.

SafeTensors errdefer fix:
- shards_mmapped counter skipped cleanup of shards at indices past the
  counter. Pre-initialize all shard entries so errdefer iterates the
  full array.

KV cache double-free guards:
- PagedKvCache.freeBlock: added debug-mode per-block scan of free list
  to catch double-free (previously only checked full-list overflow).
- TieredKvCache.freeBlockInner: same double-free guard pattern.
- block_allocator: added errdefer to free allocated block on realloc
  failure.

Tiered KV cache:
- Eviction scanned only index ranges [0..vram_count] and
  [vram_count..vram+ram_count], missing blocks that migrated between
  tiers. Now scans all blocks.
- SSD free block promotion allocated backing memory even for fresh
  blocks never spilled to SSD. Now only reads from SSD when
  ssd_offset is set.
- Free list capacity was tier-specific, but blocks migrate. Now
  reserves total_blocks capacity for each tier's free list.
- preadAll treated negative pread return (I/O error) same as EOF.
  Now returns error.ReadError on negative, only breaks on zero.
- Discrete GPU transfer path had src==dst pointer bug. Added assert
  until Plan 04 adds separate host/device pointer tracking.

Grammar enhancements:
- Added char_not_star, char_not_plus, char_not_opt element types for
  quantified negated character classes ([^x]*, [^x]+, [^x]?).
- Escape sequences inside character classes: \n, \t, \r, \\,
  \], \- now handled in both single chars and range endpoints.
- JSON grammar: added \uXXXX hex escape support (new 'hex' rule),
  scientific notation for numbers (e/E with optional sign).
- Token acceptance: grammar completion mid-token now rejects remaining
  bytes (prevents extra content after grammar is satisfied).
- BPE prefix stripping: handle decoded space-prefixed tokens (" word")
  in addition to Latin1 prefix bytes.
- acceptToken: break on completion to avoid processing past grammar end.

Other fixes:
- Accelerate vDSP_dotpr: wrong C types (c_int → c_long/c_ulong) for
  ARM64 macOS ABI.
- Format dataByteLen: floor division truncated non-block-aligned tensor
  sizes. Changed to ceiling division matching gguf.zig tensorBytes.
- Calibrate readCalFile: unchecked multiplication overflow on huge
  dimensions. Now uses std.math.mul with overflow detection. Added
  truncated file checks on all readPositionalAll calls.
- Display: buffer overflow on narrow terminals (saturating subtraction),
  rope_theta NaN/overflow guard, JSON serialization NaN guard.
- Image resize: validates source buffer size against dimensions, rejects
  zero dimensions.
- Readline: leftover buffer overflow protection on oversized input.
- WebGPU: replaced magic number array sizes with named constants, added
  assertions for bounds validation.
- Chat template: tool result messages include tool_call_id; added
  Gemma4 unified and DeepSeek4 templates to fuzz test.
- Recipes: added Qwen2, DeepSeek V2, GLM-4 tuning presets.
- expert_cache: hardcoded page size replaced with std.heap.page_size_min.

Tests added:
- softmax: all-negative-infinity and partial-infinity distributions
- calibrate: overflow-inducing header dimensions
- image: undersized source buffer, zero dimensions
- eval: scoreCase correctness and forward-failure handling
- kvcache: freeBlock double-free guard
- grammar: rule count updated for new hex rule

+++ src/backend/accelerate.zig
@@ -20,7 +20,7 @@ const CblasTrans: c_int = 112;
 // Import Accelerate BLAS functions via C ABI (resolved at link time)
 extern "c" fn cblas_sgemm(order: c_int, transa: c_int, transb: c_int, m: c_int, n: c_int, k: c_int, alpha: f32, a: [*]const f32, lda: c_int, b: [*]const f32, ldb: c_int, beta: f32, c_ptr: [*]f32, ldc: c_int) void;
 extern "c" fn cblas_sdot(n: c_int, x: [*]const f32, incx: c_int, y: [*]const f32, incy: c_int) f32;
-extern "c" fn vDSP_dotpr(a: [*]const f32, ia: c_int, b: [*]const f32, ib: c_int, c_ptr: *f32, n: c_int) void;
+extern "c" fn vDSP_dotpr(a: [*]const f32, ia: c_long, b: [*]const f32, ib: c_long, c_ptr: *f32, n: c_ulong) void;
 
 /// Matrix multiply: C[m×n] = A[m×k] × B[k×n] (row-major, B transposed).
 /// Uses AMX on M1+ for ~4× speedup over NEON.
